# Fix #1687 - export gather_tree: improved performance ...