... by not calling blender_object.children, which is slow : takes O(len(bpy.data.objects)) operations.

Related to issue https://github.com/KhronosGroup/glTF-Blender-IO/issues/1687

Instead, I build a `dict` containing parent/child relations, which is used during the traversal.

I measured time spent before/after the patch, using b3d 3.3-alpha+master.7f4ee97b9ef9 ; I see improvement from 2.05sec to 0.06sec - about 40x faster in this case.


## before patch
```
16:40:43 | INFO: Draco mesh compression is available, use library at C:\nllbin\blender-3.3.0-alpha+master.7f4ee97b9ef9-windows.amd64-release\3.3\python\lib\site-packages\extern_draco.dll
16:40:43 | INFO: Starting glTF 2.0 export
16:40:45 | INFO: Extracting primitive: Cube.1643
16:40:45 | INFO: Primitives created: 1
16:40:45 | INFO: Finished glTF 2.0 export in 2.051008939743042 s
```

## timing after patch, b3d 3.3-alpha+master.7f4ee97b9ef9
```
16:43:00 | INFO: Draco mesh compression is available, use library at C:\nllbin\blender-3.3.0-alpha+master.7f4ee97b9ef9-windows.amd64-release\3.3\python\lib\site-packages\extern_draco.dll
16:43:00 | INFO: Starting glTF 2.0 export
16:43:00 | INFO: Extracting primitive: Cube.1643
16:43:00 | INFO: Primitives created: 1
16:43:00 | INFO: Finished glTF 2.0 export in 0.06299805641174316 s
```