@julienduroure 2022-08-05
Looks good to me.
Any reason why you pass blender_children as parameter in recursive function instead of creating it as VExportTree attribute and accessing it using self.blender_children?