@rotoglup 2022-08-05
No reason except me being reluctant to add attributes when I don't fully
understand code and objects lifetimes.

I'm on vacation for now, if you'd like me to adjust, I can do so at the end
of the month.

Thanks

Le ven. 5 août 2022 à 15:46, Julien Duroure ***@***.***> a
écrit :

> Looks good to me.
> Any reason why you pass blender_children as parameter in recursive
> function instead of creating it as VExportTree attribute and accessing it
> using self.blender_children?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/KhronosGroup/glTF-Blender-IO/pull/1688#issuecomment-1206477174>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAHCDEXRFHMQSNQJ2G4PZJTVXULK3ANCNFSM54EFUMGQ>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>