@julienduroure 2022-08-05
Hello,
Lifetime is a good point. VExportTree instance will survive during all export, when blender_children is useful only during vtree construction.
I think I will commit it as is.
Thanks again, and enjoy your holidays!